+++ Libplanet/Store/Trie/ITrie.cs
@@ -1,4 +1,5 @@
 #nullable enable
+using System.Collections.Generic;
 using System.Security.Cryptography;
 using Bencodex.Types;
 
@@ -25,7 +26,7 @@ namespace Libplanet.Store.Trie
         /// Stores the <paramref name="value"/> to the
         /// node corresponding to given <paramref name="key"/>.
         /// </summary>
-        /// <param name="key">An index to look with <see cref="TryGet"/> after.</param>
+        /// <param name="key">The unique key to associate with the <paramref name="value"/>.</param>
         /// <param name="value">The value to store.</param>
         /// <exception cref="System.ArgumentNullException">Thrown when the given
         /// <paramref name="value"/> is <c>null</c>.</exception>
@@ -33,14 +34,13 @@ namespace Libplanet.Store.Trie
         ITrie Set(in KeyBytes key, IValue value);
 
         /// <summary>
-        /// Gets the value stored with <paramref name="key"/> in <see cref="Set"/>.
+        /// Gets the values stored with <paramref name="keys"/> in <see cref="Set"/>.
         /// </summary>
-        /// <param name="key">The key used in <see cref="Set"/> to store a value.</param>
-        /// <param name="value">The reference pointer to be set the value stored.</param>
-        /// <returns>If there is a value corresponded to <paramref name="key"/>,
-        /// set <paramref name="value"/> to it and return true. If not, set <paramref name="value"/>
-        /// to null and return false.</returns>
-        bool TryGet(in KeyBytes key, out IValue? value);
+        /// <param name="keys">The keys used in <see cref="Set"/> to store a value.</param>
+        /// <returns>The values associated to the specified <paramref name="keys"/>.  The associated
+        /// values are ordered in the same way to the corresponding <paramref name="keys"/>.  Absent
+        /// values are represented as <see langword="null"/>.</returns>
+        IReadOnlyList<IValue?> Get(IReadOnlyList<KeyBytes> keys);
 
         /// <summary>
         /// Cleans up and stores the <see cref="ITrie"/> in storage.